Refactor views to create new modules Add functions to models
