+++ quiz/utils/helpers.py
@@ -1,0 +1,2 @@
+def getFormattedScore(score, total):
+    return f'{score}/{total}'
